Add Authorization HTTP header check
Add an optional authorization string when registering the Webhook.  If
used, discard any incoming request that does not have a matching
Authorization header.  Otherwise, keep the same behaviour as before.
Log a warning if an Authorization header is received but none was
expected.

This is particularly useful when the callback URL is made public
(e.g. LAVA job definition) and when using an encrypted communication.
The Authorization header would typically be a token stored both in the
Jenkins secrets and in the remote end, and never be transferred in
clear.  Some application specific mechanism will be used to enable the
remote end to pick the correct token, for example by first sending a
token identifier when setting up the remote callback.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>